+++ chat-challenge/src/ui/organisms/screens/ChatScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './ChatScreen'
